[Floss] Delay calling powered callback until HCI enabled
There have been inconsistencies in the UI where if btadapterd
crashes, and we try to turn on bluetooth, the UI will show
bluetooth as turned on successfully. This is because when the UI
starts the bluetooth manager, there's no real indication of
whether or not it succeeded, thus we should delay calling the
success callback until we get HCI enabled.

Bug: b/231249514, b/228905413
Change-Id: I4dd037ef23ad7b5567efb5458496d64e10abae29
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3656851
Commit-Queue: Katherine Lai <laikatherine@chromium.org>
Reviewed-by: Abhishek Pandit-Subedi <abhishekpandit@google.com>
Cr-Commit-Position: refs/heads/main@{#1010161}