[Chrome App Deprecation] Adding 'launch anyways' button
This change adds a 'launch anyways' option to chrome apps deprecation
dialogs, as well as testing to ensure this launches the app in the
correct container.

Launch anyways is implemented by providing the dialog with a callback,
which is implemented by app_launcher_handler.cc, which has all of the
required information already to launch the given extension in the
correct context.

See the bug for screenshots, and this doc for strings review:
https://docs.google.com/document/d/1V7cVxiSHp1chAYdQ753MqJYpfIF8k0ZZpHIu6l4tgHw/edit

Bug: 1323370
Change-Id: I66c0d21c9dbe176e15fbec55510da0628721f787
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3630762
Commit-Queue: Daniel Murphy <dmurph@chromium.org>
Reviewed-by: Phillis Tang <phillis@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1010172}