Serve cached prefetched responses.
When the navigation to a URL that was prefetched starts, the prefetch is
moved into a map keyed by just the URL, and the cookie copy process
starts. The navigation interceptor then checks just this new map for a
prefetch to use. If there is a prefetch with a cached response, it
intercepts the navigation and creates a URL loader to load the
prefetched response.

The following files were copied from
chrome/browser/prefetch/prefetch_proxy/ with only minimal changes:
 - prefetch_proxy_cookie_listener* -> prefetch_cookie_listener*
 - prefetch_proxy_from_string_url_loader* ->
   prefetch_from_string_url_loader*
 - prefetch_proxy_url_loader_interceptor* ->
   prefetch_url_loader_interceptor*. There are considerably more changes
   compared to the other files listed above.

Bug: 1299059
Change-Id: If5c9405dfda2be60da35e642cc75d69210688730
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3657953
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: Ryan Sturm <ryansturm@chromium.org>
Commit-Queue: Max Curran <curranmax@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1010206}