Rename and refactor isFragmentNavigation to indicate it's only valid for primary main frame nav
The way this value was computed made it meaningless for subframe and
non-primary frame navigations. We rename it to better indicate its
semantics. Fortunately, all callers were already checking for
isInPrimaryMainFrame before calling this.

There were also a few redundant uses of this where the caller had
already checked that the navigation was not same document, so it
couldn't be a fragment navigation. We remove these redundant calls.

Bug: 1196715
Change-Id: Ibcf0aaf57c566d5cd99c2987b3437c87e1038c33
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3684455
Reviewed-by: Ayman Almadhoun <ayman@chromium.org>
Reviewed-by: Zhiyuan Cai <zhiyuancai@chromium.org>
Reviewed-by: Bo Liu <boliu@chromium.org>
Commit-Queue: Kevin McNee <mcnee@chromium.org>
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1010232}