Rework AnimationBetweenRows() to support reversing animations.
When dragging around an app item back and forth between rows, the
row change animation would not properly reverse the animation.

This CL rewrites AnimationBetweenRows() and moves the functionality
into a new class AppsGridRowChangeAnimator. Now interrupting the
animation with another row change will perfectly reverse the currently
running animation when necessary. This is accomplished by creating
a layer copy of the row changing item view. The layer copy is animated
out of the grid towards the screen edge. The item view is animated into
the new row position from the screen edge. When this row change
animation is reversed, the layer copy and the item view switch places
and are animated to the new correct target bounds.

For a drag going across multiple rows, many items can have a row change
animation happening. For this case, the layer copies are stored in a map
called `row_change_layers_`, which associates each row change item with
its copied layer.

Also add a unittest for the case where the row change animation is
reversed, to check that the view animates in the correct positions.

Bug: 1318259
Change-Id: I180053530c305178587888408bdd18a034bbf2ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3603237
Commit-Queue: Matthew Mourgos <mmourgos@chromium.org>
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1010247}