Add policy for Renderer App Container.
This policy will allow Renderer App Container to be disabled.

Setting this policy to Disabled will override platform and field-trial
configuration and force Renderer App Container to be disabled.

Leaving this policy unset, or Enabled, will defer the control of
whether or not Renderer App Container is enabled to platform, Chrome
version, or field trial configuration.

This CL queries the policy on the UI thread in the constructor
for RendererSandboxedProcessLauncherDelegateWin but the value
is not used until the renderer is launched on Launcher thread.

A test is added to verify this behavior. In addition, some
cleanup is done of enums in ContentBrowserClient.

BUG=1328867

Change-Id: I98986646e8b27cffcd6c4567b9baea5fe6994e85
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3680781
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Commit-Queue: Will Harris <wfh@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1010283}