OOFs and the initial column balancing pass
We lay out the OOFs during the initial column balancing pass once they
reach the CB rather than the fragmentation context root. As a result,
we may get different results in a non-initial column balancing pass.

Thus, miss the cache if the initial column balancing pass between the
new and old constraint spaces are different, and there is an OOF in
the fragmentainer subtree.

Bug: 1329877,1329424
Change-Id: I32e771579f333e268f87610548c72b2c6f3be4d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3682172
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1010296}