Eliminate base::LowerCaseEqualsASCII in /components/dom_distiller
This CL is a behavior-preserving refactor. It replaces calls to
base::LowerCaseEqualsASCII with base::EqualsCaseInsensitiveASCII.

We are deprecating base::LowerCaseEqualsASCII because its interface is a
mild footgun, in that it requires the second parameter to be lowercase.

See the design doc for more details.
https://docs.google.com/document/d/18zWV0zN4oAqMZpond1Ut0QviHUeHCSwCo08mqThwHkQ/edit?usp=sharing

This CL was uploaded by git cl split.

R=wychen@chromium.org

Bug: 1326529
Change-Id: I5ad185e3416c1ab3981f91fe5a5fd32d2f9b3b35
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3655707
Auto-Submit: Dan McArdle <dmcardle@chromium.org>
Commit-Queue: Wei-Yin Chen <wychen@chromium.org>
Reviewed-by: Wei-Yin Chen <wychen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1010302}