AXPosition: Split out boundary behavior from initial move behavior
In an upcoming CL, we need a behavior that would map to
kStopAtLastAnchorBoundaryOrIfAlreadyAtBoundary - in attempts to add that
functionality, the ||'s and &&'s were difficult to comprehend both
reading and writing the code.

This CL decouples the two concepts into separate enums, which are
passed in to AXPosition's methods via a combined struct.

This is a straight refactor with no behavior change. Follow up changes
will add tests for the new combinations, and UIA code will opt into the
required behavior in another follow up CL.

Change-Id: I543d1de177ad7ec3ea796ee1e8cd7316138fa728
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3671747
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Reviewed-by: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Commit-Queue: Daniel Libby <dlibby@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1010313}