NotificationRevamp: Fix the convert animation in multiple displays.
The convert from single to group animation in secondary display is
broken (do a slide in instead of just size change) since we use the
global MessageCenter to update pop-up animations (which is changed in
the primary display). This CL add a setup to all the displays before any
conversion are started.

Change-Id: Ibeea6e10f31c23922fa87789554a14158b911725
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3669153
Commit-Queue: Andre Le <leandre@chromium.org>
Reviewed-by: Ahmed Mehfooz <amehfooz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1010314}