[Extensions cleanup] Update site access items based on the new section
Instead of comparing the new site interaction with the current section's
site interaction, we can compare directly the sections. Basically we get
the new section and reorder/remove/reinsert based on the current section.

This allows us to remove `site_interaction` from SiteAccessSection,
which is wanted because user site permissions may affect the items in
the section.

Bug: 1263310
Change-Id: I30c87aca3d1e209de11e96144965738825ed4057
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3685373
Commit-Queue: Emilia Paz <emiliapaz@chromium.org>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1010375}