Eliminate base::LowerCaseEqualsASCII in /net/cert
This CL is a behavior-preserving refactor. It replaces calls to
base::LowerCaseEqualsASCII with base::EqualsCaseInsensitiveASCII.

We are deprecating base::LowerCaseEqualsASCII because its interface is a
mild footgun, in that it requires the second parameter to be lowercase.

See the design doc for more details.
https://docs.google.com/document/d/18zWV0zN4oAqMZpond1Ut0QviHUeHCSwCo08mqThwHkQ/edit?usp=sharing

This CL was uploaded by git cl split.

R=mattm@chromium.org

Bug: 1326529
Change-Id: Ibfcc9d748a3667b7de801976129aa11212eaba05
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3658088
Auto-Submit: Dan McArdle <dmcardle@chromium.org>
Reviewed-by: Matt Mueller <mattm@chromium.org>
Commit-Queue: Matt Mueller <mattm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1010416}