Refactor LinuxInputMethodContext::UpdateFocus.
This is preparation to remove context_simple_ from InputMethodAuralinux.
This introduces LinuxInputMethodContext::UpdateFocus, and
replaces LinuxInputMethodContext::{Focus,Blur} by the new method.

Importantly, Focus,Blur handling logic are distributed to
InputMethodContextImplGtk and WaylandInputMethodContext.
It's known that it is not working properly for wayland(Lacros),
and this is also the preparation of the fix. With this series of
refactoring, we can fix the behavior in wayland without impacting
GTK.

This is just a refactoring, and no behaivor change should be observed
out side from InputMethodAuralinux.

BUG=1331183
TEST=Ran locally.

Change-Id: Ib7f32eb86ad49aa0814801c53c99474c674acd27
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3685950
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Reviewed-by: Darren Shen <shend@chromium.org>
Commit-Queue: Hidehiko Abe <hidehiko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1010419}