Set terminal tab background color to avoid flashes
When we load a new tab, we can immediately set the correct background
color by inspecting the URL 'settings_profile' param, or by using the
same color as the opener tab.

If neither are found, we fallback to the color set for 'default'
profile, or the default color.

Bug: 1330425
Change-Id: I43657c421cc5f579291609b164402e6c33949332
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3685869
Commit-Queue: Joel Hockey <joelhockey@chromium.org>
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1010431}