WrappedSkImage - Unify support check between GL/Vulkan
WrappedSkImage earlier checked for usage support based on whether
SkiaRenderer was used with GL or Vulkan. Now, we don't have a special
path for GL anymore and that has been removed as part of CopySubTexture
for GL removal in RasterDecoder. As a result, GL and Vulkan work the
same and so we unify it here in the IsSupported check.

Also, fix the RasterDecoder unit tests which earlier tested with
SharedImageGLRepresentation instead of SharedImageSkiaRepresentation.

Change-Id: Ib08d1a776877209b151a0f234e3535d57d6a1794
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3647740
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Saifuddin Hitawala <hitawala@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1010436}