Filter out uninstalling, stub and non-exposed web apps in FindAppWithUrlInScope()
This CL is a rewrite of
https://chromium-review.googlesource.com/c/chromium/src/+/3648150
which enabled FindAppWithUrlInScope() to search for uninstalling web
apps but also enabled it to search for stub and non-WebAppExposed()
web apps.

This broke FindAppWithUrlInScope() when System Web Apps are included
in the web app database of the Lacros profile. This condition is
currently hit during Lacros profile migration.

FindAppWithUrlInScope() should not include web apps outside of
GetApps() so this CL reverts that change and updates the
AppBannerManagerDesktop::OnWebAppWillBeUninstalled()
callsite to use IsUrlInAppScope() instead which still works on
uninstalling web apps.

Bug: 1330432, 1177421
Change-Id: I3cc0867805935e4aaf57f322a07501e53a13bcdc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3681720
Reviewed-by: Eric Willigers <ericwilligers@chromium.org>
Commit-Queue: Alan Cutter <alancutter@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Reviewed-by: Yuta Hijikata <ythjkt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1010454}