Simplify ProfileHelper for testing.
This removes two methods from ProfileHelper.
- GetLockScreenIncognitoProfile: No users. Just remove.
- GetProfileByUserIdHashForTest: Used only by tests.

The latter one is more complicated and the main focus of this CL.
Despite of its name, it "Creates" a Profile instance synchronously,
and in many uses, this creation is what tests want.

Thus, this CL replaces many of them by profiles::testing::CreateProfileSync,
which is a test utility designed for the purpose.

In path_util_unittest.cc, it is used just a global getter and not the
main test target, so instead use members to keep Profile instances.

BUG=1325210
TEST=Ran bots.

Change-Id: I9cecd55766e8bd6b72285e73b9a786a566bad093
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3675746
Reviewed-by: David Trainor <dtrainor@chromium.org>
Commit-Queue: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1010458}