Define new field types for phone number pre- and suffix.
For example, the American phone number (650) 234-5678 maps to the
existing types as follows:
- PHONE_HOME_CITY_CODE = 650
- PHONE_HOME_NUMBER = 2345678
With the new types we additionally get (once they are implemented):
- PHONE_HOME_NUMBER_PREFIX = 234
- PHONE_HOME_NUMBER_SUFFIX = 5678

These types correspond to the HTML autocomplete types tel-local-prefix
and tel-local-suffix, which currently map to PHONE_HOME_NUMBER.
This mapping is not updated yet, as currently special case handling
ensures that they are treated correctly for US numbers. Once
the new types are implemented, this mapping will be updated.

Bug: 1311937
Change-Id: Ifb7ebc2a21270e08ddb29ac10e8ea137e421d6f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3667605
Reviewed-by: Christoph Schwering <schwering@google.com>
Commit-Queue: Florian Leimgruber <fleimgruber@google.com>
Cr-Commit-Position: refs/heads/main@{#1010511}