float: Add education nudge under the maximize button p1.
This nudge will let users know there is a hover to float/maximize/etc.
on the maximize/restore button. The nudge works similar to system
nudges, but does not use the framework since invoking it is quite
different, and it is anchored.

Specs have not arrived, so will be done later.

Bug: 1329233
Test: manual

Change-Id: Iaa9f3f3e160fd9fffac27d625fb71baf0a041e03
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3669446
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1010562}