Fix Flaky Test: QuotaManagerImplTest.RetrieveBucketsTable
It has been determined that the likely culprit for the flake resulting
from this test is the 'ModifyStageKeyAndNotify' method. This method takes in
a callback that is triggered after the usage update is triggered.

Previously, this callback is set to base::DoNothing(). With this change,
the ModifyStorageKeyAndNotify method has been modified to receive an
optional base::OnceClosure instead. As a result,
the test itself has been updated to use a run_loop to ensure that the
rest of the test will continue running only after the usage has been
updated - which should resolve the flake.

Bug: 1329201
Change-Id: I38667d3699ffc7e8a922d2823bd4392a4337d6ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3687358
Reviewed-by: Ayu Ishii <ayui@chromium.org>
Commit-Queue: Christine Smith <christinesm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1010589}