[iOS] Unescape original URL from error page URL
The error page URL are capturing the original URL (the one which failed)
and escape its value.
When the error page helper is returning the failed URL based on an error
page URL in failedNavigationURLFromErrorPageFileURL, the URL should be
un-escaped to ensure that the same URL is returned.

This is important as the URL is sometimes compared to another URL, which
fails if there is an encoded character.

Fixed: 1332538
Change-Id: I3c6573d56b8331a5a24636eb5b55b3f50203782e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3687792
Commit-Queue: Ali Juma <ajuma@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1010598}