ash: Fix launcher appearing in screenshots triggered by assistant
When the assistant hears the phrase "take a screenshot" it closes the
launcher and takes a screenshot. The launcher must close immediately
(without animation) otherwise it appears in the screenshot, which is
probably not what the user intended.

Fix by wiring up AppListControllerImpl::ShouldDismissImmediately()
to AppListBubbleView for the clamshell bubble launcher.

Also add a check for feature ProductivityLauncher inside of
ShouldDismissImmediately() to make it clear some of the code only
runs with peeking launcher (and therefore can be deleted when
peeking launcher is removed).

Bug: 1329425
Test: added to ash_unittests
Change-Id: I8733f30aede2c89ca8729785d945ad295cd9e4d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3688331
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Commit-Queue: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1010608}