Keep partitioned cookies whose partition key is same-site with their domain
When we convert partitioned cookies to unpartitioned when a site opts
out of the CHIPS origin trial, right now we keep the most recently used
partitioned cookie.

After this change, we will prioritize any partitioned cookie which has a
partition key which is same-site with the cookie's domain.

Bug: 1331114
Change-Id: I638327aeab030536e284f01bbce3df0f9895b0cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3686539
Commit-Queue: Dylan Cutler <dylancutler@google.com>
Reviewed-by: Steven Bingler <bingler@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1010610}