personalization: Fix Google Photos wallpaper staleness check on login.
When the user logs in w/ a Google Photos wallpaper set, we are supposed
to check if the photo still exists in the user's Google Photos library.
If not, the wallpaper should be reset back to default.

Currently, the check is failing because the auth token needed to confirm
(in-)existence of the photo is failing to be fetched. This is due to a
timing issue as the primary account access token fetcher is not yet
primed to handle a one-shot request for an access token during wallpaper
initialization on login.

To fix, change token fetcher modes to wait until login completes.

Bug: b:234456652
Change-Id: Ie559ff75c01e205f12e7ac3259e0133adeb4cd4e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3687995
Reviewed-by: Colin Kincaid <ckincaid@chromium.org>
Reviewed-by: Jeffrey Young <cowmoo@chromium.org>
Commit-Queue: David Black <dmblack@google.com>
Cr-Commit-Position: refs/heads/main@{#1010690}