Remove DCHECK that there is at least one stage after event dispatch
This is a removal of the DCHECK added in r1009001. This is not an exact
revert as the early return the revert adds would interfere with the
correct recording of the trace event.

The reason for the revert is r1009001 has caused some flakiness in
tests.

Bug: 1330903
Change-Id: If8a625dac7f1957161a77f6b4ef488e249fef016
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3686237
Commit-Queue: Mohsen Izadi <mohsen@chromium.org>
Reviewed-by: Behdad Bakhshinategh <behdadb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1010696}