DPWA: fix UAF on launch --uninstall-app-id
`WebAppUninstallDialogDelegateView` is a `DialogDelegate` and is owned by its widget. It was
triggering a callback on `OnDialogCanceled`.
The callback was passed from `WebAppUiManagerImpl` to release a
`ScopedKeepAlive`, which could trigger a browser shutdown and further
attempts to close the dialog widget, which races with the widget &
delegate's destruction managed by views framework.

The fix removes the redundant and too-early callback call on
`OnDialogCanceled` and only call it on destruction.

Bug: 1330289
Change-Id: I660c0af08a460ce6cedace8d0871dcb925ad9d76
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3688264
Commit-Queue: Phillis Tang <phillis@chromium.org>
Reviewed-by: Chase Phillips <cmp@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1010715}